Moved typeMismatchErrorMessage_shouldBeDisplayedWhenItOccurs to CatchFormDesignExceptionsTest
